Refactored resource out of structure, css updates
